requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
Automat==22.10.0
axios==0.3.0
beautifulsoup4==4.11.1
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colored==1.4.3
coloredlogs==15.0.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.2.0
ddddocr==1.4.7
ddt==1.5.0
dill==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
future==0.18.2
googletrans==4.0.0rc1
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.6.0
incremental==22.10.0
iniconfig==1.1.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
logzero==1.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
minium==1.2.8
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mysqlclient==2.1.1
numpy==1.25.0
onnxruntime==1.15.1
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
parsel==1.8.1
pathos==0.3.0
Pillow==9.2.0
playwright==1.33.0
pluggy==1.0.0
pox==0.3.2
ppft==1.7.6.6
prettytable==3.6.0
Protego==0.2.1
protobuf==4.23.3
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyecharts==2.0.2
pyee==9.0.4
Pygments==2.15.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
pytest==7.1.2
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==10.2
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
rich==13.3.4
runstats==2.0.0
schedule==1.2.0
Scrapy==2.8.0
scrapy-zyte-api==0.8.2
selenium==4.3.0
service-identity==21.1.0
simple-tornado==0.2.2
simplejson==3.18.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
sympy==1.12
tenacity==8.2.2
tidevice==0.4.10
tldextract==3.4.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.3.0
tzdata==2023.3
undetected-chromedriver==3.4.7
urllib3==1.26.10
w3lib==2.1.1
wcwidth==0.2.6
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.1.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
zyte-api==0.4.5
